refactor: Move `MainScreen` to a separate file
