+++ src/main/kotlin/ui/App.kt
@@ -24,12 +24,10 @@ import com.halilibo.richtext.ui.material.SetupMaterialRichText
 
 import kotlinx.coroutines.launch
 
-import org.ossreviewtoolkit.model.Identifier
 import org.ossreviewtoolkit.utils.common.titlecase
 import org.ossreviewtoolkit.workbench.composables.FileDialog
 import org.ossreviewtoolkit.workbench.model.OrtApiState
 import org.ossreviewtoolkit.workbench.navigation.Navigation
-import org.ossreviewtoolkit.workbench.navigation.OldScreen
 import org.ossreviewtoolkit.workbench.navigation.rememberNavigationController
 import org.ossreviewtoolkit.workbench.theme.OrtWorkbenchTheme
 import org.ossreviewtoolkit.workbench.ui.dependencies.Dependencies
@@ -40,7 +38,6 @@ import org.ossreviewtoolkit.workbench.ui.settings.Settings
 import org.ossreviewtoolkit.workbench.ui.summary.Summary
 import org.ossreviewtoolkit.workbench.ui.violations.Violations
 import org.ossreviewtoolkit.workbench.ui.vulnerabilities.Vulnerabilities
-import org.ossreviewtoolkit.workbench.utils.MaterialIcon
 
 @Composable
 fun App(controller: WorkbenchController) {
@@ -68,17 +65,6 @@ fun App(controller: WorkbenchController) {
             }
         }
     }
-}
-
-sealed class MainScreen(val name: String, val icon: MaterialIcon) : OldScreen {
-    object Dependencies : MainScreen("Dependencies", MaterialIcon.ACCOUNT_TREE)
-    object Issues : MainScreen("Issues", MaterialIcon.BUG_REPORT)
-    class PackageDetails(val pkg: Identifier) : MainScreen("Package Details", MaterialIcon.INVENTORY)
-    object Packages : MainScreen("Packages", MaterialIcon.INVENTORY)
-    object Settings : MainScreen("Settings", MaterialIcon.SETTINGS)
-    object Summary : MainScreen("Summary", MaterialIcon.ASSESSMENT)
-    object RuleViolations : MainScreen("Rule Violations", MaterialIcon.GAVEL)
-    object Vulnerabilities : MainScreen("Vulnerabilities", MaterialIcon.LOCK_OPEN)
 }
 
 @Composable
